a good batch of fixes including MMC2 support
added the frameCounter IRQ
a bunch of PPU bugs:
added a delay to help synchronize PPU and CPU (not ideal)
fix wrong clearing of sprit hit etc
don't handle leftmost block if PPUMASK says so

Added Mapper MMC2 which allows us to play Punch Out!
todo: clear MMC2 a bit
todo: this game uses DMC so we can finally debug/fix it

The NMI fix allows us to play bomber man!

todo: might be time to start allowing save games?

+++ nes/common/rom.go
@@ -24,9 +24,9 @@ func (r *Rom) Read16(addr uint16) uint16 {
 	return uint16(r.Read8(addr)) | uint16(r.Read8(addr+1))<<8
 }
 func (r *Rom) Write8(addr uint16, val uint8) {
-	r.Write8w(addr, val)
+	r.Write8w(uint32(addr), val)
 }
-func (r *Rom) Write8w(addr uint16, val uint8) {
+func (r *Rom) Write8w(addr uint32, val uint8) {
 	if r.writable {
 		r.rom[addr] = val
 	} else {
